+++ lib/l10n/cat_l10n.dart
@@ -128,8 +128,6 @@ extension CatL10n on S {
         return stageAdolescent;
       case 'adult':
         return stageAdult;
-      case 'senior':
-        return stageSenior;
       default:
         return stageId;
     }